requirements.txt
streamlit==1.28.1
openai==1.3.5
PyPDF2==3.0.1
python-docx==0.8.11
openpyxl==3.10.10
pandas==2.0.3